// frontend/src/api/generated/models/productIssue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Dekopen API
 * Authenticated tenant and engine API boundary.
 * OpenAPI spec version: 1.0.0
 */
import type { ProductIssueParams } from "./productIssueParams";
import type { ProductIssueSeverityEnum } from "./productIssueSeverityEnum";

export interface ProductIssue {
  code: string;
  severity: ProductIssueSeverityEnum;
  target: string;
  params: ProductIssueParams;
}